+++ src/linearmixedmodel.jl
@@ -122,7 +122,7 @@ function LinearMixedModel(
                 wtz,
                 convert(LowerTriangular{Float64,Matrix{Float64}}, x.λ),
                 x.inds,
-                convert(SparseMatrixCSC{T,Int32}, x.adjA),
+                convert(SparseMatrixCSC{T,Int64}, x.adjA),
                 convert(Matrix{T}, x.scratch),
             )
             push!(reterms, x)
@@ -1156,10 +1156,10 @@ Base.show(io::IO, m::LinearMixedModel) = Base.show(io, MIME"text/plain"(), m)
     _coord(A::AbstractMatrix)
 
 Return the positions and values of the nonzeros in `A` as a
-`NamedTuple{(:i, :j, :v), Tuple{Vector{Int32}, Vector{Int32}, Vector{Float64}}}`
+`NamedTuple{(:i, :j, :v), Tuple{Vector{Int64}, Vector{Int64}, Vector{Float64}}}`
 """
 function _coord(A::Diagonal)
-    return (i=Int32.(axes(A, 1)), j=Int32.(axes(A, 2)), v=A.diag)
+    return (i=Int64.(axes(A, 1)), j=Int64.(axes(A, 2)), v=A.diag)
 end
 
 function _coord(A::UniformBlockDiagonal)
@@ -1167,13 +1167,13 @@ function _coord(A::UniformBlockDiagonal)
     r, c, k = size(dat)
     blk = repeat(r .* (0:(k - 1)); inner=r * c)
     return (
-        i=Int32.(repeat(1:r; outer=c * k) .+ blk),
-        j=Int32.(repeat(1:c; inner=r, outer=k) .+ blk),
+        i=Int64.(repeat(1:r; outer=c * k) .+ blk),
+        j=Int64.(repeat(1:c; inner=r, outer=k) .+ blk),
         v=vec(dat),
     )
 end
 
-function _coord(A::SparseMatrixCSC{T,Int32}) where {T}
+function _coord(A::SparseMatrixCSC{T,Int64}) where {T}
     rv = rowvals(A)
     cv = similar(rv)
     for j in axes(A, 2), k in nzrange(A, j)
@@ -1187,8 +1187,8 @@ _coord(A::BlockedSparse) = _coord(A.cscmat)
 function _coord(A::Matrix)
     m, n = size(A)
     return (
-        i=Int32.(repeat(axes(A, 1); outer=n)),
-        j=Int32.(repeat(axes(A, 2); inner=m)),
+        i=Int64.(repeat(axes(A, 1); outer=n)),
+        j=Int64.(repeat(axes(A, 2); inner=m)),
         v=vec(A),
     )
 end
@@ -1196,7 +1196,7 @@ end
 """
     sparseL(m::LinearMixedModel; fname::Symbol=first(fnames(m)), full::Bool=false)
 
-Return the lower Cholesky factor `L` as a `SparseMatrix{T,Int32}`.
+Return the lower Cholesky factor `L` as a `SparseMatrix{T,Int64}`.
 
 `full` indicates whether the parts of `L` associated with the fixed-effects and response
 are to be included.
@@ -1214,12 +1214,12 @@ function sparseL(
     end
     blks = sblk:(length(reterms) + full)
     rowoffset, coloffset = 0, 0
-    val = (i=Int32[], j=Int32[], v=T[])
+    val = (i=Int64[], j=Int64[], v=T[])
     for i in blks, j in first(blks):i
         Lblk = L[block(i, j)]
         cblk = _coord(Lblk)
-        append!(val.i, cblk.i .+ Int32(rowoffset))
-        append!(val.j, cblk.j .+ Int32(coloffset))
+        append!(val.i, cblk.i .+ Int64(rowoffset))
+        append!(val.j, cblk.j .+ Int64(coloffset))
         append!(val.v, cblk.v)
         if i == j
             coloffset = 0
